Correção para model.find_same = nil Linha 28

+++ app/mixins/csv_model_delta.rb
@@ -25,7 +25,7 @@ module CSVModelDelta
           model.save
 
         elsif model.deletable?
-          model.find_same.destroy
+          model.find_same.destroy unless model.find_same.nil?
 
         elsif model.updatable?
           model_to_update = model.find_same
